commit(1): add 2 fails and their in main.js
